+++ processor/src/lib.rs
@@ -950,3 +950,27 @@ pub(crate) fn add_error_ctx_to_external_error(
         },
     }
 }
+
+/// Reads (start_addr, end_addr) tuple from the specified elements of the operand stack (
+/// without modifying the state of the stack), and verifies that memory range is valid.
+pub fn get_mem_addr_range(
+    process: &ProcessState,
+    start_idx: usize,
+    end_idx: usize,
+) -> Result<(u32, u32), MemoryError> {
+    let start_addr = process.get_stack_item(start_idx).as_int();
+    let end_addr = process.get_stack_item(end_idx).as_int();
+
+    if start_addr > u32::MAX as u64 {
+        return Err(MemoryError::address_out_of_bounds(start_addr, &()));
+    }
+    if end_addr > u32::MAX as u64 {
+        return Err(MemoryError::address_out_of_bounds(end_addr, &()));
+    }
+
+    if start_addr > end_addr {
+        return Err(MemoryError::InvalidMemoryRange { start_addr, end_addr });
+    }
+
+    Ok((start_addr as u32, end_addr as u32))
+}