+++ stdlib/src/handlers/mod.rs
@@ -3,6 +3,7 @@ use miden_core::Felt;
 pub mod falcon_div;
 pub mod keccak256;
 pub mod smt_peek;
+pub mod sorted_array;
 pub mod u64_div;
 
 // HELPER FUNCTIONS